requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
plotly==5.24.1
pytest==8.3.3
pip install openpyxl
